RDKPREINTG-13472: Backend APIS for Modules, Features, Parameter
Reason for change: TDK Revamp: Backend APIS for Modules, Features, Parameter
Test Procedure: with TDK TM
Risks: None
Priority: P1

+++ tdkservice/tdkservice/src/main/java/com/rdkm/tdkservice/repository/DeviceRepositroy.java
@@ -29,7 +29,7 @@ import com.rdkm.tdkservice.model.BoxType;
 import com.rdkm.tdkservice.model.Device;
 
 @Repository
-public interface DeviceRepositroy extends JpaRepository<Device, Long> {
+public interface DeviceRepositroy extends JpaRepository<Device, Integer> {
 	/**
 	 * This method is used to check if a Device exists by its stbIp.
 	 *
